refactor(modo-email): extract mailer factory functions into factory.rs
Move mailer() and mailer_with() out of lib.rs into a dedicated factory
module, following the file-per-responsibility convention.
