+++ modo-email/src/lib.rs
@@ -21,6 +21,7 @@
 //! ```
 
 mod config;
+mod factory;
 mod mailer;
 mod message;
 pub mod template;
@@ -37,33 +38,6 @@ pub use config::ResendConfig;
 #[cfg(feature = "smtp")]
 pub use config::SmtpConfig;
 
+pub use factory::{mailer, mailer_with};
 pub use template::filesystem::FilesystemProvider;
 pub use template::layout::LayoutEngine;
-
-use std::sync::Arc;
-
-/// Create a [`Mailer`] using [`FilesystemProvider`] and the transport configured in `config`.
-///
-/// This is the standard entry point. Templates are loaded from `config.templates_path`.
-pub fn mailer(config: &EmailConfig) -> Result<Mailer, modo::Error> {
-    let provider = Arc::new(FilesystemProvider::new(&config.templates_path));
-    mailer_with(config, provider)
-}
-
-/// Create a [`Mailer`] with a custom [`TemplateProvider`].
-///
-/// Use this when you want to load templates from a database, cache, or any
-/// source other than the filesystem.
-pub fn mailer_with(
-    config: &EmailConfig,
-    provider: Arc<dyn TemplateProvider>,
-) -> Result<Mailer, modo::Error> {
-    let transport = transport::transport(config)?;
-    let layout = Arc::new(LayoutEngine::new(&config.templates_path));
-    let sender = SenderProfile {
-        from_name: config.default_from_name.clone(),
-        from_email: config.default_from_email.clone(),
-        reply_to: config.default_reply_to.clone(),
-    };
-    Ok(Mailer::new(transport, provider, sender, layout))
-}
